pyproject.toml
[tool.poetry]
name = "bible-com-scraper"
version = "0.1.0"
description = "Scrape bibles from bible.com"
authors = ["Tage Johansson <frans.tage@gmail.com>"]
readme = "README.md"
packages = [{include = "bible_com_scraper.py"}]

[tool.poetry.scripts]
bible_com_scraper = "bible_com_scraper:main"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.11.1"
httpx = "^0.23.3"
progressbar2 = "^4.2.0"


[tool.poetry.group.dev.dependencies]
types-beautifulsoup4 = "^4.11.6.4"
types-requests = "^2.28.11.8"
mypy = "^0.991"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
